Fixed some instructions(cast) + bad formatted instruction jump

+++ Vm/load_index.c
@@ -6,7 +6,7 @@
 /*   By: ahamouda <ahamouda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/29 14:42:59 by ahamouda          #+#    #+#             */
-/*   Updated: 2016/09/02 06:28:28 by ahamouda         ###   ########.fr       */
+/*   Updated: 2016/09/05 23:34:42 by ahamouda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void		load_index(t_vm_data *arena, t_proc *process)
 	tmp_two = get_n_param_value(arena, process, 2, &error);
 	reg = get_n_reg_param_value(arena, process, 3, &error);
 	tmp += tmp_two;
-	tmp = get_param_value(arena, (process->pc + (tmp % IDX_MOD)), REG_SIZE);
+	tmp = get_param_value(arena, (process->pc + (((short)tmp) % IDX_MOD)), REG_SIZE);
 	move_pc_from_format(arena, process);
 	if (error)
 		return (display_ldi_instruction(process, arena, 0));
